+++ wechat/src/main/java/com/gallery/ui/wechat/WeChatUiResult.kt
@@ -2,10 +2,9 @@ package com.gallery.ui.wechat
 
 object WeChatUiResult {
     const val GALLERY_WE_CHAT_RESULT_FULL_IMAGE = "galleryWeChatResultFullImage"
-    const val GALLERY_WE_CHAT_RESULT_PREV_IMAGE = "galleryWeChatResultPrev"
-    const val GALLERY_WE_CHAT_RESULT_PREV_IDS = "galleryWeChatResultIds"
     const val GALLERY_WE_CHAT_VIDEO_DURATION = "galleryWeChatVideoDuration"
     const val GALLERY_WE_CHAT_VIDEO_DES = "galleryWeChatVideoDes"
+    const val GALLERY_WE_CHAT_VIDEO_ALL = "galleryWeChatVideoAll"
 
     const val GALLERY_WE_CHAT_ALL_VIDEO_PARENT: Long = (-112).toLong()
 }