+++ wechat/src/main/java/com/gallery/ui/wechat/activity/GalleryWeChatPrevActivity.kt
@@ -6,8 +6,6 @@ import android.os.Bundle
 import android.view.View
 import android.widget.FrameLayout
 import androidx.core.content.ContextCompat
-import androidx.kotlin.expand.os.getBooleanExpand
-import androidx.kotlin.expand.os.getSerializableOrDefault
 import com.gallery.core.GalleryBundle
 import com.gallery.core.delegate.prevFragment
 import com.gallery.scan.ScanEntity
@@ -15,12 +13,13 @@ import com.gallery.scan.types.isGifExpand
 import com.gallery.scan.types.isVideoExpand
 import com.gallery.ui.activity.PrevBaseActivity
 import com.gallery.ui.adapter.GalleryFinderAdapter
-import com.gallery.ui.wechat.R
-import com.gallery.ui.wechat.WeChatUiResult
+import com.gallery.ui.wechat.*
+import com.gallery.ui.wechat.WeChatPrevArgs.Companion.weChatPrevArgsOrDefault
+import com.gallery.ui.wechat.WeChatPrevSaveArgs.Companion.putArgs
+import com.gallery.ui.wechat.WeChatPrevSaveArgs.Companion.weChatPrevSaveArgs
 import com.gallery.ui.wechat.adapter.WeChatPrevSelectAdapter
 import com.gallery.ui.wechat.engine.displayGalleryPrev
 import com.gallery.ui.wechat.engine.displayGalleryPrevSelect
-import com.gallery.ui.wechat.obtain
 import kotlinx.android.synthetic.main.gallery_activity_wechat_prev.*
 
 @SuppressLint("SetTextI18n")
@@ -30,9 +29,7 @@ class GalleryWeChatPrevActivity : PrevBaseActivity(R.layout.gallery_activity_wec
         get() = R.id.preWeChatFragment
 
     private val selectAdapter: WeChatPrevSelectAdapter = WeChatPrevSelectAdapter(this)
-    private val isPrev: Boolean by lazy { uiGapConfig.getBooleanExpand(WeChatUiResult.GALLERY_WE_CHAT_RESULT_PREV_IMAGE) }
-    private val videoDuration: Int by lazy { uiGapConfig.getInt(WeChatUiResult.GALLERY_WE_CHAT_VIDEO_DURATION) }
-    private val fullImageSelect: Boolean by lazy { uiGapConfig.getBooleanExpand(WeChatUiResult.GALLERY_WE_CHAT_RESULT_FULL_IMAGE) }
+    private val weChatPrevArgs: WeChatPrevArgs by lazy { uiGapConfig.weChatPrevArgsOrDefault }
     private val idList: ArrayList<Long> = ArrayList()
 
     private fun onUpdateVideoTip(scanEntity: ScanEntity) {
@@ -43,9 +40,9 @@ class GalleryWeChatPrevActivity : PrevBaseActivity(R.layout.gallery_activity_wec
             prevWeChatSelect.setTextColor(ContextCompat.getColor(this, android.R.color.white))
             return
         }
-        if (scanEntity.duration > videoDuration) {
+        if (scanEntity.duration > weChatPrevArgs.videoDuration) {
             galleryPrevVideoTip.visibility = View.VISIBLE
-            galleryPrevVideoTip.text = getString(R.string.gallery_select_video_prev_error).format(videoDuration / 1000 / 60)
+            galleryPrevVideoTip.text = getString(R.string.gallery_select_video_prev_error).format(weChatPrevArgs.videoDuration / 1000 / 60)
             prevWeChatSelect.setTextColor(ContextCompat.getColor(this, R.color.color_999999))
             prevWeChatToolbarSend.isEnabled = false
             prevWeChatSelect.isEnabled = false
@@ -60,18 +57,18 @@ class GalleryWeChatPrevActivity : PrevBaseActivity(R.layout.gallery_activity_wec
 
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
-        outState.putSerializable(WeChatUiResult.GALLERY_WE_CHAT_RESULT_PREV_IDS, idList)
+        WeChatPrevSaveArgs(idList).putArgs(outState)
     }
 
     override fun onCreate(savedInstanceState: Bundle?) {
+        uiPrevArgs
         super.onCreate(savedInstanceState)
         obtain(uiConfig)
         idList.clear()
-        @Suppress("UNCHECKED_CAST")
-        idList.addAll(savedInstanceState.getSerializableOrDefault(WeChatUiResult.GALLERY_WE_CHAT_RESULT_PREV_IDS, ArrayList<Long>()) as ArrayList<Long>)
+        idList.addAll(savedInstanceState?.weChatPrevSaveArgs?.ids ?: arrayListOf())
         prevWeChatToolbarBack.setOnClickListener { onGalleryFinish() }
         prevWeChatToolbarText.text = "%s / %s".format(0, galleryConfig.multipleMaxCount)
-        prevWeChatFullImage.isChecked = fullImageSelect
+        prevWeChatFullImage.isChecked = weChatPrevArgs.fullImageSelect
         galleryPrevList.adapter = selectAdapter
         prevWeChatToolbarSend.isEnabled = true
         prevWeChatToolbarSend.setOnClickListener {
@@ -88,7 +85,7 @@ class GalleryWeChatPrevActivity : PrevBaseActivity(R.layout.gallery_activity_wec
     }
 
     override fun onGalleryFinderThumbnails(finderEntity: ScanEntity, container: FrameLayout) {
-        container.displayGalleryPrevSelect(finderEntity, idList, isPrev)
+        container.displayGalleryPrevSelect(finderEntity, idList, weChatPrevArgs.isPrev)
     }
 
     override fun onDisplayGalleryPrev(galleryEntity: ScanEntity, container: FrameLayout) {
@@ -111,7 +108,7 @@ class GalleryWeChatPrevActivity : PrevBaseActivity(R.layout.gallery_activity_wec
         prevWeChatSelect.setOnClickListener { prevFragment.checkBoxClick(it) }
         galleryPrevList.visibility = if (prevFragment.selectEntities.isEmpty()) View.GONE else View.VISIBLE
         galleryPrevListLine.visibility = if (prevFragment.selectEntities.isEmpty()) View.GONE else View.VISIBLE
-        selectAdapter.updateSelect(prevFragment.selectEntities)
+        selectAdapter.updateSelect(if (weChatPrevArgs.isPrev) uiPrevArgs.prevArgs.selectList else prevFragment.selectEntities)
         onUpdateVideoTip(prevFragment.currentItem)
         prevFragment.view?.findViewById<View>(R.id.preCheckBox)?.visibility = View.GONE
     }
@@ -119,7 +116,7 @@ class GalleryWeChatPrevActivity : PrevBaseActivity(R.layout.gallery_activity_wec
     override fun onChangedCheckBox() {
         val currentItem = prevFragment.currentItem
         prevWeChatToolbarSend.text = uiConfig.selectText + if (prevFragment.selectEmpty) "" else "(${prevFragment.selectCount}/${galleryConfig.multipleMaxCount})"
-        if (isPrev) {
+        if (weChatPrevArgs.isPrev) {
             if (!currentItem.isSelected) {
                 idList.add(currentItem.id)
             } else {